Add ChargeGas method to Runtime
It will be used in the current effort of setting sane values for gas.
Gas is split into compute gas and storage gas to allow better analisys
of computation time in relation to gas.
Both gas types are latter combined into one and this is charged for the
execution.

Virtual gas is added for observing how changes to gas numbers affect
the balance without affecting the stateroot of the chain.

Signed-off-by: Jakub Sztandera <kubuxu@protocol.ai>