WIP: refactor state
WIP: remove cbor gen code

WIP: constructor logic and sector expiration logic
I think we're going to need to rework this to avoid loading everything into
memory all at once. At the very least, we're creating bitfields that are too
large.

WIP remove NewSectorExpirations
We can't actually use these. We'll need to load new sectors at the proving
period, regardless of what we do, so we can load power.

fix popSectorExpirations and attach to the state.

attach PopExpiredSectors to the state
This way, we can export it and test it.