Avoid reloading AMT for every sector loaded (#649)
* Avoid reloading AMT for every sector loaded

The AMT caches on read internally so we can save quite a few loads by _not_
throwing the array away after every sector info load. This is especially
important when loading ranges of sequential sectors (which should be a pretty
common case).

* Avoid reloading sector info AMT for each proof

Same as before, re-loading the array for a likely sequential set of sector IDs
is wasteful.