Change the WPoSt commit randomness lookback to match the proof challenge epoch
This change gets rid of the variable ChainCommitRand parameter and fixes it to
be `DeadlineOpen - WPoStChallengeLookback`.

This makes it easy for lotus to extract randomness from the chain at the
beginning of the challenge window such that:

1. The randomness is stable and won't change due to reverts. The WPoStChallengeLookback should ensure this.
2. No WPoSt message will become _invalid_ before the challenge window ends.

Before, if lotus picked a commit epoch that was too far in the past, it ran the
risk of not being able to submit the proof in time. However, if it picked a
recent epoch, it ran the risk of a revert changing the chain.