harden the deadline liveness test
Make sure the deadline is _really_ dead before skipping it during the deadline cron.

We check:

* There are live sectors (the obvious test).
* There are no proven partitions (shouldn't be possible, but might as well check).
* The partitions haven't changed since last time (dead for at least one proving period).
* There are no optimistic proofs in the snapshot (dead for at least one proving period).

This means all "dead" partitions should be "normalized" (partitions = partitionsSnapshot, etc.).
If, for example, a miner decides to "compact" a dead partition, the deadline cron will run for
that partition exactly once in order to re-normalize it.