Implements a new PreCommitSectorBatch method, as specified by FIP-0008 (discussion)
The non-batched PreCommitSector method remains, but should be considered deprecated.
The implementation delegates internally to the batched method.
This consumes less gas than before this change (due to removal of redundant checks in
the batched method) but is slightly less efficient at runtime than a strictly-optimized
singleton method could be. I expect it to be removed in the future.