Raise AddressedSectorsMax from 10,000 to 25,000.
This number bounds the amount of memory used or state loaded in a number of operations
such as declaring faults. A critical need is to limit the number of sectors that may
be addressed by a single WindowPoSt sumbission to ensure it can be safely disputed
later if fraudulent.

25000 sectors permits submitting (and disputing) up to 10 partitions at once
for both 32GiB (25,000/2,349 ~= 10.6) and 64GiB (25,000/2,300 ~= 10.8) sectors.

Empirically, disputing 10 partitions fits within the block gas limit >= 3 times over.